refactor(transformer/class-properties): rename file (#8036)
Pure refactor. Rename file to make filename more descriptive of what it does.
