+++ crates/oxc_transformer/src/es2022/class_properties/mod.rs
@@ -175,18 +175,18 @@
 //!
 //! Implementation is split into several files:
 //!
-//! * `mod.rs`:                Setup and visitor.
-//! * `class.rs`:              Transform of class body.
-//! * `prop_decl.rs`:          Transform of property declarations (instance and static).
-//! * `constructor.rs`:        Insertion of property initializers into class constructor.
-//! * `instance_prop_init.rs`: Transform of instance property initializers.
-//! * `static_prop_init.rs`:   Transform of static property initializers and static blocks.
-//! * `computed_key.rs`:       Transform of property/method computed keys.
-//! * `private_field.rs`:      Transform of private fields (`this.#prop`).
-//! * `super.rs`:              Transform `super` expressions.
-//! * `class_details.rs`:      Structures containing details of classes and private properties.
-//! * `class_bindings.rs`:     Structure containing bindings for class name and temp var.
-//! * `utils.rs`:              Utility functions.
+//! * `mod.rs`:                        Setup and visitor.
+//! * `class.rs`:                      Transform of class body.
+//! * `prop_decl.rs`:                  Transform of property declarations (instance and static).
+//! * `constructor.rs`:                Insertion of property initializers into class constructor.
+//! * `instance_prop_init.rs`:         Transform of instance property initializers.
+//! * `static_block_and_prop_init.rs`: Transform of static property initializers and static blocks.
+//! * `computed_key.rs`:               Transform of property/method computed keys.
+//! * `private_field.rs`:              Transform of private fields (`this.#prop`).
+//! * `super.rs`:                      Transform `super` expressions.
+//! * `class_details.rs`:              Structures containing details of classes and private properties.
+//! * `class_bindings.rs`:             Structure containing bindings for class name and temp var.
+//! * `utils.rs`:                      Utility functions.
 //!
 //! ## References
 //!
@@ -215,7 +215,7 @@ mod constructor;
 mod instance_prop_init;
 mod private_field;
 mod prop_decl;
-mod static_prop_init;
+mod static_block_and_prop_init;
 mod supers;
 mod utils;
 use class_bindings::ClassBindings;
